requirements.txt
biopython
numpy
matplotlib
pandas
pyemma
pyrosetta
pytraj
scikit-learn